% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.MAST.r
\name{write_results.MAST}
\alias{write_results.MAST}
\title{Write MAST model fitting results to file(s)}
\usage{
\code{
write_results.MAST(lfc_result,
                   write_full_results=TRUE,
                   write_sigGenes=TRUE, thresholds.adj.P.Val=0.05,
                   write_topGenes=TRUE, n_topGenes=100,
                   file_prefix, test_id)}
}
\arguments{
\item{lfc_result}{The fit results, generally the output of \code{MASTtools::getModelResults.MAST}}

\item{write_full_results}{logical, whether to output the full results data frame to a file. Defaults to TRUE.}

\item{write_sigGenes}{logical, whether to output a list of the genes with adjusted p-values below a threshold. Defaults to TRUE.}

\item{thresholds.adj.P.Val}{numeric, a vector of adjusted p-value thresholds to use for gene significance cutoffs. One file is generated for each threshold value. Defaults to 0.05. Ignored if \code{write_sigGenes == FALSE}}

\item{write_topGenes}{logical, whether to output a sorted list of the genes with strongest significance. Defaults to TRUE.}

\item{n_topGenes}{numeric, a vector of numbers of most significant genes to output. One file is generated for each threshold value. Defaults to 100. Ignored if \code{write_topGenes == FALSE}}

\item{file_prefix}{character, the path to the file location. "MAST_..." will be appended to it.}

\item{test_id}{character, the name of the model test, to be included in the filename.}
}
\description{
Output file(s) with results from MAST model fitting
}
